fixed #3
fixed issue #3: Getter gets a Float, but should be a Double

+++ src/main/java/me/grison/jtoml/Getter.java
@@ -31,11 +31,11 @@ public interface Getter {
     Integer getInt(String key);
 
     /**
-     * Get the float.
+     * Get the double value.
      * @param key the key where the object is located
      * @return the float located at the given key.
      */
-    Float getFloat(String key);
+    Double getDouble(String key);
 
     /**
      * Get the date (as Calendar)
